Django_REST_backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.3.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.5
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
wrapt==1.14.1
